feat:Change wav data to float64 and remove jit export

+++ util.py
@@ -89,7 +89,7 @@ def librosa_mel(x):
     mel = librosa.feature.melspectrogram(y=x, sr=16000, n_fft=512, win_length=400, hop_length=160,
                                          window=scipy.signal.windows.hamming, n_mels=80, fmin=20,
                                          fmax=7600)
-    return mel.astype(numpy.float32)
+    return mel
 
 
 def torch_mel(x):
